pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "python-security-scanner"
dynamic = ["version"]
description = "A professional security vulnerability scanner for Python applications"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["security", "vulnerability", "scanner", "static-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.0.0",
    "colorama>=0.4.4",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "isort>=5.0",
    "flake8>=3.9",
    "mypy>=0.900",
    "pre-commit>=2.0",
]

[project.urls]
Homepage = "https://github.com/your-username/python-security-scanner"
Documentation = "https://github.com/your-username/python-security-scanner/docs"
Repository = "https://github.com/your-username/python-security-scanner"
Issues = "https://github.com/your-username/python-security-scanner/issues"

[project.scripts]
security-scanner = "security_scanner.cli:main"

[tool.setuptools_scm]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=security_scanner --cov-report=html --cov-report=term-missing"
testpaths = ["tests"]